Expert list

+++ database/migrations/2016_10_10_190139_experts.php
@@ -17,7 +17,7 @@ class Experts extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('email')->unique();
-            $table->mediumInteger('mobile')->nullable();
+            $table->bigInteger('mobile')->nullable();
             $table->timestamps();
         });
     }
